fix: Add onFilterChange handler to ProductFilterWrapper component
