+++ src/app/page.tsx
@@ -1,4 +1,3 @@
-"use client";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -15,7 +14,6 @@ export default function HomePage() {
   return (
     <main className="container mx-auto px-4 py-8">
       <h1 className="text-3xl font-bold mb-8">Welcome to Our Marketing App</h1>
-      {/* Add home page content */}
     </main>
   );
 }
