+++ src/app/products/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from "next";
 import { Product } from "@/types";
 import ProductGrid from "@/components/ProductGrid";
-import ProductFilter from "@/components/ProductFilter";
+import ProductFilterWrapper from "@/components/ProductFilterWrapper";
 
 export const metadata: Metadata = {
   title: "Products | Marketing App",
@@ -22,21 +22,12 @@ async function getProducts(): Promise<Product[]> {
 export default async function ProductsPage() {
   const products = await getProducts();
 
-  const handleFilterChange = (filters: {
-    category: string;
-    minPrice: number | "";
-    maxPrice: number | "";
-  }) => {
-    // Filter implementation will come in next feature
-    console.log("Filters:", filters);
-  };
-
   return (
     <main className="container mx-auto px-4 py-8">
       <h1 className="text-3xl font-bold mb-8">Our Products</h1>
       <div className="flex flex-col md:flex-row gap-8">
         <aside className="w-full md:w-64">
-          <ProductFilter onFilterChange={handleFilterChange} />
+          <ProductFilterWrapper />
         </aside>
         <ProductGrid products={products} />
       </div>
